+++ src/main/java/edu/illinois/library/cantaloupe/auth/Authorizer.java
@@ -8,6 +8,17 @@ import java.io.IOException;
 public interface Authorizer {
 
     /**
+     * Authorizes a request early in the request cycle, before any information
+     * is available about the source image.
+     *
+     * @return Instance reflecting the authorization status of the request.
+     */
+    AuthInfo preAuthorize() throws IOException;
+
+    /**
+     * Authorizes a request after full information is available about the
+     * source image.
+     *
      * @return Instance reflecting the authorization status of the request.
      */
     AuthInfo authorize() throws IOException;