+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/v1/ImageResource.java
@@ -68,6 +68,11 @@ public class ImageResource extends IIIF1Resource {
             }
 
             @Override
+            public boolean preAuthorize() throws Exception {
+                return ImageResource.this.preAuthorize();
+            }
+
+            @Override
             public boolean authorize() throws Exception {
                 return ImageResource.this.authorize();
             }