+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/v2/ImageResource.java
@@ -75,6 +75,11 @@ public class ImageResource extends IIIF2Resource {
             }
 
             @Override
+            public boolean preAuthorize() throws Exception {
+                return ImageResource.this.preAuthorize();
+            }
+
+            @Override
             public boolean authorize() throws Exception {
                 return ImageResource.this.authorize();
             }